cleanup: Drop obsolete comment about HsConDetails
HsConDetails used to have an argument representing the type of the
tyargs in a list:

    data HsConDetails tyarg arg rec
      = PrefixCon [tyarg] [arg]

This datatype was shared across 3 synonyms: HsConPatDetails,
HsConDeclH98Details, HsPatSynDetails. In the latter two cases, `tyarg`
was instanced to `Void` meaning the list was always empty for these
cases.

In 7b84c58867edca57a45945a20a9391724db6d9e4, this was refactored such
that HsConDetails no longer needs a type of tyargs by construction. The
first case now represents the type arguments in the args type itself,
with something like:

    ConPat "MkE" [InvisP tp1, InvisP tp2, p1, p2]

So the deleted comment really is just obsolete.

Fixes #26461